+++ tightdb-java-test/src/test/java/com/tightdb/TableSpecEqualityTest.java
@@ -50,14 +50,14 @@ public class TableSpecEqualityTest {
         TableSpec spec1 = new TableSpec();
         spec1.addColumn(ColumnType.STRING, "foo");
         spec1.addColumn(ColumnType.TABLE, "bar");
-        spec1.getSubtableSpec(1).addColumn(ColumnType.LONG, "x");
+        spec1.getSubtableSpec(1).addColumn(ColumnType.INTEGER, "x");
         spec1.getSubtableSpec(1).addColumn(ColumnType.TABLE, "sub");
         spec1.getSubtableSpec(1).getSubtableSpec(1).addColumn(ColumnType.BOOLEAN, "b");
 
         TableSpec spec2 = new TableSpec();
         spec2.addColumn(ColumnType.STRING, "foo");
         spec2.addColumn(ColumnType.TABLE, "bar");
-        spec2.getSubtableSpec(1).addColumn(ColumnType.LONG, "x");
+        spec2.getSubtableSpec(1).addColumn(ColumnType.INTEGER, "x");
         spec2.getSubtableSpec(1).addColumn(ColumnType.TABLE, "sub");
         spec2.getSubtableSpec(1).getSubtableSpec(1).addColumn(ColumnType.BOOLEAN, "b");
 
@@ -69,14 +69,14 @@ public class TableSpecEqualityTest {
         TableSpec spec1 = new TableSpec();
         spec1.addColumn(ColumnType.STRING, "foo");
         spec1.addColumn(ColumnType.TABLE, "bar");
-        spec1.getSubtableSpec(1).addColumn(ColumnType.LONG, "x");
+        spec1.getSubtableSpec(1).addColumn(ColumnType.INTEGER, "x");
         spec1.getSubtableSpec(1).addColumn(ColumnType.TABLE, "sub");
         spec1.getSubtableSpec(1).getSubtableSpec(1).addColumn(ColumnType.BOOLEAN, "b");
 
         TableSpec spec2 = new TableSpec();
         spec2.addColumn(ColumnType.STRING, "foo");
         spec2.addColumn(ColumnType.TABLE, "bar");
-        spec2.getSubtableSpec(1).addColumn(ColumnType.LONG, "x");
+        spec2.getSubtableSpec(1).addColumn(ColumnType.INTEGER, "x");
         spec2.getSubtableSpec(1).addColumn(ColumnType.TABLE, "sub2");
         spec2.getSubtableSpec(1).getSubtableSpec(1).addColumn(ColumnType.BOOLEAN, "b");
